backend/long_memory/requirements.txt
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.3.7
Pillow==9.2.0
django-cors-headers==3.13.0
python-dateutil==1.4
